split dev ui demos into separate pages
