+++ .github/instructions/eslint-rules.instructions.md
@@ -6,3 +6,6 @@ The following commands should be used to test the code:
 
 - Tests: `moon run eslint-rules:test`
 - Prettier: `moon run root:prettierCheck`
+
+These can be run from any directory, there's no need to `cd` to a particular
+directory.
