+++ pinyinly.code-workspace
@@ -8,6 +8,10 @@
     "[git-commit]": {
       "editor.rulers": [80],
     },
+    "github.copilot.chat.agent.terminal.allowList": {
+      "moon run app:lint -- --fix": true,
+      "/^moon run [\\w-]+:[\\w-]+$/": true,
+    },
     "editor.defaultFormatter": "esbenp.prettier-vscode",
     "typescript.tsdk": "node_modules/typescript/lib",
     "editor.codeActionsOnSave": {
