+++ projects/app/src/app/dev/ui/_demos/NewSprout.tsx
@@ -1,0 +1,5 @@
+import { NewSprout } from "@/client/ui/NewSprout";
+
+export default () => {
+  return <NewSprout className="h-[320px] w-[415px]" />;
+};
