+++ projects/app/src/app/dev/ui/_demos/QuizDeckHanziToPinyinQuestion.tsx
@@ -1,0 +1,28 @@
+import { QuizDeckHanziToPinyinQuestion } from "@/client/ui/QuizDeckHanziToPinyinQuestion";
+import { Use } from "@/client/ui/Use";
+import { hanziWordToPinyinQuestionOrThrow } from "@/data/questions/hanziWordToPinyin";
+import { hanziWordToPinyin } from "@/data/skills";
+import { buildHanziWord } from "@/dictionary/dictionary";
+
+export default () => {
+  const skill = hanziWordToPinyin(buildHanziWord(`你好`, `hello`));
+  const questionPromise = hanziWordToPinyinQuestionOrThrow(skill);
+
+  return (
+    <Use
+      promise={questionPromise}
+      render={(question) => (
+        <QuizDeckHanziToPinyinQuestion
+          noAutoFocus
+          onNext={() => {
+            console.log(`onNext()`);
+          }}
+          onRating={() => {
+            console.log(`onRating()`);
+          }}
+          question={question}
+        />
+      )}
+    />
+  );
+};
