+++ projects/app/src/app/dev/ui/_demos/WikiHanziWordModal.tsx
@@ -1,0 +1,16 @@
+import { WikiHanziWordModal } from "@/client/ui/WikiHanziWordModal";
+import { View } from "react-native";
+
+export default () => {
+  return (
+    <>
+      <View>
+        <WikiHanziWordModal
+          devUiSnapshotMode
+          hanziWord={`你好:hello`}
+          onDismiss={() => null}
+        />
+      </View>
+    </>
+  );
+};
